refactor: Update ResourceStoreStorageAdapterFactory to use ExpiringStorage
