+++ config/presets/identity/solid-identity-provider.json
@@ -12,15 +12,15 @@
       "@id": "urn:solid-server:default:IdpStorageAdapterFactory",
       "@type": "ClientWebIdFetchingStorageAdapterFactory",
       "ClientWebIdFetchingStorageAdapterFactory:_wrappedAdapterFactory": {
-        "@type": "ResourceStoreStorageAdapterFactory",
-        "ResourceStoreStorageAdapterFactory:_args_baseUrl": {
+        "@type": "ExpiringStorageAdapterFactory",
+        "ExpiringStorageAdapterFactory:_args_baseUrl": {
           "@id": "urn:solid-server:default:variable:baseUrl"
         },
-        "ResourceStoreStorageAdapterFactory:_args_storagePathname": "/idp/oidc",
-        "ResourceStoreStorageAdapterFactory:_args_store": {
-          "@type": "PassthroughKeyValueStore",
-          "PassthroughKeyValueStore:_store": {
-            "@id": "urn:solid-server:default:ResourceStore"
+        "ExpiringStorageAdapterFactory:_args_storagePathname": "/idp/oidc",
+        "ExpiringStorageAdapterFactory:_args_storage": {
+          "@type": "WrappedExpiringStorage",
+          "WrappedExpiringStorage:_source": {
+            "@id": "urn:solid-server:default:IdpStorage"
           }
         }
       }
