+++ src/index.ts
@@ -110,11 +110,11 @@ export * from './identity/IdentityProviderHttpHandler';
 export * from './identity/configuration/KeyGeneratingIdpConfigurationGenerator';
 
 // SolidIdentityProvider/Storage
+export * from './identity/storage/ClientWebIdFetchingStorageAdapterFactory';
+export * from './identity/storage/ExpiringStorageAdapterFactory';
 export * from './identity/storage/KeyValueStore';
 export * from './identity/storage/PassthroughKeyValueStore';
 export * from './identity/storage/StorageAdapterFactory';
-export * from './identity/storage/ResourceStoreStorageAdapterFactory';
-export * from './identity/storage/ClientWebIdFetchingStorageAdapterFactory';
 
 // SolidIdentityProvider/InteractionPolicies/Util
 export * from './identity/interaction/util/BasicIssuerReferenceWebIdOwnershipValidator';
