+++ test/unit/identity/storage/ExpiringStorageAdapterFactory.test.ts
@@ -1,0 +1,9 @@
+import {
+  ExpiringStorageAdapterFactory,
+} from '../../../../src/identity/storage/ExpiringStorageAdapterFactory';
+
+describe('ExpiringStorageAdapterFactory', (): void => {
+  it('satisfies a trivial use case.', async(): Promise<void> => {
+    expect(ExpiringStorageAdapterFactory).toBeDefined();
+  });
+});
